test-requirements.txt
coverage>=4.0
flake8>=2.2.0
mock>=1.0.1
PyHamcrest>=1.8.5
pytest>=2.8.0
pytest-cov>=2.1.0
pytest-xdist>=1.13.1
pytest-capturelog>=0.7
xmltodict>=0.9.2
cobertura-clover-transform>=1.1.4
fasteners>=0.12.0